requirements.txt
locustio==0.8.1
statsd==3.1
